requirements.txt
flask
pytest
numpy
pillow